# back/docs/openapi.yaml
openapi: 3.0.0
paths:
...
  /specialities:
    post:
      operationId: SpecialitiesController_create
      parameters: []
      requestBody:
        required: true
        description: New speciality payload
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: Mage
              required:
                - name
              additionalProperties: false
      responses:
        '201':
          description: Speciality created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: number
                    example: 5
                  name:
                    type: string
                    example: Mage
      security:
        - bearer: []
      tags:
        - Specialities
    get:
      operationId: SpecialitiesController_findAll
      parameters: []
      responses:
        '200':
          description: List of specialities
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: number
                      example: 1
                    name:
                      type: string
                      example: Warrior
      security:
        - bearer: []
      tags:
        - Specialities
  /specialities/{id}:
    get:
      operationId: SpecialitiesController_findOne
      parameters:
        - name: id
          required: true
          in: path
          description: Speciality ID
          schema:
            example: 5
            type: number
      responses:
        '200':
          description: Speciality by id
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: number
                    example: 5
                  name:
                    type: string
                    example: Mage
      security:
        - bearer: []
      tags:
        - Specialities
    patch:
      operationId: SpecialitiesController_update
      parameters:
        - name: id
          required: true
          in: path
          description: Speciality ID
          schema:
            example: 5
            type: number
      requestBody:
        required: true
        description: Fields to update (partial)
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: Battlemage
              additionalProperties: false
      responses:
        '200':
          description: Updated speciality
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: number
                    example: 5
                  name:
                    type: string
                    example: Battlemage
      security:
        - bearer: []
      tags:
        - Specialities
    delete:
      operationId: SpecialitiesController_remove
      parameters:
        - name: id
          required: true
          in: path
          description: Speciality ID
          schema:
            example: 5
            type: number
      responses:
        '200':
          description: Delete result
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: number
                    example: 5
                  deleted:
                    type: boolean
                    example: true
      security:
        - bearer: []
      tags:
        - Specialities
...
  /statuses:
    get:
      operationId: StatusesController_list
      parameters: []
      responses:
        '200':
          description: List of statuses
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: number
                      example: 1
                    name:
                      type: string
                      example: Pending
      security:
        - bearer: []
      tags:
        - Statuses
    post:
      operationId: StatusesController_create
      parameters: []
      requestBody:
        required: true
        description: New status payload
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: In Progress
              required:
                - name
              additionalProperties: false
      responses:
        '201':
          description: Status created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: number
                    example: 3
                  name:
                    type: string
                    example: In Progress
      security:
        - bearer: []
      tags:
        - Statuses
  /statuses/{id}:
    patch:
      operationId: StatusesController_update
      parameters:
        - name: id
          required: true
          in: path
          description: Status ID
          schema:
            example: 3
            type: number
      requestBody:
        required: true
        description: Fields to update (partial)
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: Completed
              additionalProperties: false
      responses:
        '200':
          description: Updated status
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: number
                    example: 3
                  name:
                    type: string
                    example: Completed
      security:
        - bearer: []
      tags:
        - Statuses
    delete:
      operationId: StatusesController_delete
      parameters:
        - name: id
          required: true
          in: path
          description: Status ID
          schema:
            example: 3
            type: number
      responses:
        '200':
          description: Delete result
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: number
                    example: 3
                  deleted:
                    type: boolean
                    example: true
      security:
        - bearer: []
      tags:
        - Statuses
...
info:
  title: API "Les kips AH"
  description: |-
    This is the API information regarding the project "Les capuches d'Opale".

    C'est un outil de gestion d'aventurier pour une guilde.
  version: 1.0.0
  contact: {}
  termsOfService: https://swagger.io/terms/
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http